Separate NPP functions to another file
